+++ resources/views/user/home.blade.php
@@ -1,0 +1,13 @@
+<x-user-layout>
+  @slot('customCSS')
+
+  @endslot
+
+  @slot('content')
+    <h1>Hello World</h1>
+  @endslot
+
+  @slot('customJS')
+
+  @endslot
+</x-user-layout>