+++ routes/user.php
@@ -1,3 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+
+Route::get('/user/home', function () {
+    return view('user/home');
+});