+++ routes/web.php
@@ -10,10 +10,6 @@ Route::get('/login', function () {
     return view('login');
 });
 
-Route::get('/blank', function () {
-    return view('blank');
-});
-
 // This will Includes all the other route files.
 require base_path('routes/admin.php');
 require base_path('routes/adminbackend.php');